Build the c module in a custom temp dir
MSVC has a path length limit of 260 bytes, which cannot be increased.
When building a device-detection-cxx native extension, setuptools places
object files in a very long location, that MSVC cannot open due to said
limit. To work around this limitation, we create another alternative
directory, with a shorter path compared to the default, and set it as
TEMP dir.

Co-authored-by: Bohdan <sniperik@users.noreply.github.com>